refactor : revised name of service folder - 'service' -> 'apis' - reason : for understandable to anyone - applied to every files
